Updated destructors for CSVParser, DataSet, Tester
CSVParser puts instances of AttributeVector on the heap, so it should be
responsible for deleting them.

DataSet does not, however, put instances of AttributeVector on the heap,
so it should do nothing when it destructs.

Tester puts Perceptron* pointers on the heap, so it should destroy them
when it destructs.